Website infrastructure
Will update once pkgdown doesn't depend on a bunch of tidyverse packages
that are pre-CRAN release as I don't have time to see what it breaks right
now.  This means it's hard to update the website without jumping through
hoops to get the staticdocs package from the archive - the version here:
https://github.com/hadley/pkgdown/tree/67b8193f28d4ca5311d4e1049473021f2819f067
works for me.  How Tidy.